Fixed getVotesAverage() in GiocoDao.java and other improvements

+++ src/model/dao/interfaces/GiocoDaoInterface.java
@@ -11,7 +11,7 @@ public interface GiocoDaoInterface{
   public void deleteGame(Gioco gioco) throws SQLException;
   public List<Gioco> allGames() throws SQLException;
   public void deleteAllGames() throws SQLException;
-  public double getVotesAverage(Gioco gioco) throws SQLException;
+  public float getVotesAverage(Gioco gioco) throws SQLException;
   public List<Recensione> allGameReviews(Gioco gioco) throws SQLException;
   public boolean gameAlredyVotedByUser(Utente utente, Gioco gioco) throws SQLException;
 }
